Login Server
Login server for our product. Takes in user id and password and checks
it against the database. Currently doesn't encrypt password. Also checks
supervisor flag and if true it runs webui. Just opens a text file at the
moment if false (placeholder for Arcade executable).  Has a version of
the database with only LoginInstance and User. Will adjust it when I get
the full database up and running